Add attack, heal, reveal all cards effect

+++ Assets/MemoryMatch/Scripts/MainGame/CardConfigs.cs
@@ -14,13 +14,13 @@ public class CardConfig
 {
     [SerializeField] int id;
     [SerializeField] CardCategory category;
-    [SerializeField] float damage;
+    [SerializeField] int damage;
     [SerializeField] Sprite icon;
     [SerializeField] CardTarget target;
 
     public int ID => id;
     public CardCategory Category => category;
-    public float Damage => damage;
+    public int Damage => damage;
     public Sprite Icon => icon;
     public CardTarget Target => target;
 }
